+++ src/holmes-rs/src/hydro/mod.rs
@@ -1,6 +1,7 @@
 use pyo3::prelude::*;
 pub mod bucket;
 pub mod cequeau;
+pub mod crec;
 pub mod gr4j;
 pub mod utils;
 use crate::utils::register_submodule;
@@ -13,6 +14,7 @@ pub fn make_module(py: Python<'_>) -> PyResult<Bound<'_, PyModule>> {
     register_submodule(py, &m, &gr4j::make_module(py)?, "holmes_rs.hydro")?;
     register_submodule(py, &m, &bucket::make_module(py)?, "holmes_rs.hydro")?;
     register_submodule(py, &m, &cequeau::make_module(py)?, "holmes_rs.hydro")?;
+    register_submodule(py, &m, &crec::make_module(py)?, "holmes_rs.hydro")?;
     Ok(m)
 }
 
@@ -23,6 +25,7 @@ pub fn get_model(
         "gr4j" => Ok((gr4j::init, gr4j::simulate)),
         "bucket" => Ok((bucket::init, bucket::simulate)),
         "cequeau" => Ok((cequeau::init, cequeau::simulate)),
+        "crec" => Ok((crec::init, crec::simulate)),
         _ => Err(HydroError::WrongModel(model.to_string())),
     }
 }