+++ src/holmes-rs/tests/unit/hydro/mod.rs
@@ -1,3 +1,4 @@
 mod bucket_tests;
 mod cequeau_tests;
+mod crec_tests;
 mod gr4j_tests;